+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # diff-coverage-idea-plugin Changelog
 
 ## [Unreleased]
+### Changed
+- Diff Coverage run configuration instead of shortcut
+
+## [0.0.4]
 ### Fixed
 - Show Diff Coverage report notification if `diffCoverage` Gradle task failed due to violation rules
 - Don't display all packages on code coverage view with checked 'Flatten Packages'